# Measure perf of aliasing react.min.js as react