--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,7 @@
+# Webpack output
+bundle.js
+bundle.*.js
+
 # Logs
 logs
 *.log