--- a/run.sh
+++ b/run.sh
@@ -9,6 +9,8 @@ echo "Raw JS:"
 node index.js 400
 echo "Uglified: "
 node bundle.js 400
+echo "Aliasing react.min.js as react: "
+node bundle.react-min.js 400
 
 echo ""
 echo "200k"
@@ -16,3 +18,5 @@ echo "Raw JS:"
 node index.js 4000
 echo "Uglified: "
 node bundle.js 4000
+echo "Aliasing react.min.js as react: "
+node bundle.react-min.js 4000