This is a technique I've seen recommended a few times, which is lower touch than uglifying your entire codebase (including dependencies) but doesn't seem to provide the same perf gains.

On my humble 2011 Macbook Air, these are the numbers from my last run:

```
20k
Raw JS:
21.11ms
Uglified:
3.17ms
Alias React.min.js as React:
13.94ms

200k
Raw JS:
182.11ms
Uglified:
30.51ms
Alias React.min.js as React:
132.64ms
```